requirements.txt
Flask>=2.3.0
PyYAML>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
PyGithub>=1.59.0

